+++ src/main/java/de/johni0702/minecraft/bobby/Bobby.java
@@ -1,7 +1,9 @@
 package de.johni0702.minecraft.bobby;
 
 import ca.stellardrift.confabricate.Confabricate;
+import de.johni0702.minecraft.bobby.commands.UpgradeCommand;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
 import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.ModContainer;
 import net.minecraft.client.MinecraftClient;
@@ -10,6 +12,8 @@ import org.spongepowered.configurate.CommentedConfigurationNode;
 import org.spongepowered.configurate.ConfigurateException;
 import org.spongepowered.configurate.reference.ConfigurationReference;
 import org.spongepowered.configurate.reference.ValueReference;
+
+import static net.fabricmc.fabric.api.client.command.v1.ClientCommandManager.literal;
 
 public class Bobby implements ClientModInitializer {
     public static final String MOD_ID = "bobby";
@@ -33,6 +37,10 @@ public class Bobby implements ClientModInitializer {
         } catch (ConfigurateException e) {
             e.printStackTrace();
         }
+
+        ClientCommandManager.DISPATCHER.register(literal("bobby")
+                .then(literal("upgrade").executes(new UpgradeCommand()))
+        );
     }
 
     public BobbyConfig getConfig() {