+++ src/main/java/de/johni0702/minecraft/bobby/FakeChunkStorage.java
@@ -3,6 +3,7 @@ package de.johni0702.minecraft.bobby;
 import com.mojang.serialization.Codec;
 import de.johni0702.minecraft.bobby.ext.ChunkLightProviderExt;
 import de.johni0702.minecraft.bobby.mixin.LightingProviderAccessor;
+import io.netty.util.concurrent.DefaultThreadFactory;
 import net.minecraft.SharedConstants;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
@@ -13,10 +14,14 @@ import net.minecraft.nbt.NbtElement;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.nbt.NbtLongArray;
 import net.minecraft.nbt.NbtOps;
+import net.minecraft.network.MessageType;
+import net.minecraft.text.TranslatableText;
+import net.minecraft.util.Util;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.ChunkSectionPos;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.Heightmap;
 import net.minecraft.world.LightType;
 import net.minecraft.world.World;
@@ -28,19 +33,35 @@ import net.minecraft.world.chunk.ChunkSection;
 import net.minecraft.world.chunk.PalettedContainer;
 import net.minecraft.world.chunk.WorldChunk;
 import net.minecraft.world.chunk.light.LightingProvider;
+import net.minecraft.world.gen.chunk.ChunkGenerator;
+import net.minecraft.world.gen.chunk.FlatChunkGenerator;
+import net.minecraft.world.storage.StorageIoWorker;
 import net.minecraft.world.storage.VersionedChunkStorage;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.BiConsumer;
 import java.util.function.Supplier;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class FakeChunkStorage extends VersionedChunkStorage {
     private static final Logger LOGGER = LogManager.getLogger();
@@ -81,8 +102,13 @@ public class FakeChunkStorage extends VersionedChunkStorage {
         active.clear();
     }
 
+    private final Path directory;
+    private final AtomicBoolean sentUpgradeNotification = new AtomicBoolean();
+
     private FakeChunkStorage(Path directory) {
         super(directory, MinecraftClient.getInstance().getDataFixer(), false);
+
+        this.directory = directory;
     }
 
     public void save(ChunkPos pos, NbtCompound chunk) {
@@ -90,7 +116,18 @@ public class FakeChunkStorage extends VersionedChunkStorage {
     }
 
     public @Nullable NbtCompound loadTag(ChunkPos pos) throws IOException {
-        return getNbt(pos);
+        NbtCompound nbt = getNbt(pos);
+        if (nbt != null && nbt.getInt("DataVersion") != SharedConstants.getGameVersion().getSaveVersion().getId()) {
+            if (sentUpgradeNotification.compareAndSet(false, true)) {
+                MinecraftClient client = MinecraftClient.getInstance();
+                client.submit(() -> {
+                    TranslatableText text = new TranslatableText("bobby.upgrade.required");
+                    client.submit(() -> client.inGameHud.addChatMessage(MessageType.SYSTEM, text, Util.NIL_UUID));
+                });
+            }
+            return null;
+        }
+        return nbt;
     }
 
     public NbtCompound serialize(WorldChunk chunk, LightingProvider lightingProvider) {
@@ -341,6 +378,83 @@ public class FakeChunkStorage extends VersionedChunkStorage {
             }
 
             return new ChunkNibbleArray(belowBytes);
+        }
+    }
+
+    public void upgrade(RegistryKey<World> worldKey, BiConsumer<Integer, Integer> progress) throws IOException {
+        Optional<RegistryKey<Codec<? extends ChunkGenerator>>> generatorKey =
+                Optional.of(Registry.CHUNK_GENERATOR.getKey(FlatChunkGenerator.CODEC).orElseThrow());
+
+        List<ChunkPos> chunks;
+        try (Stream<Path> stream = Files.list(directory)) {
+            chunks = stream
+                    .map(Path::getFileName)
+                    .map(Path::toString)
+                    .map(Pattern.compile("^r\\.(-?[0-9]+)\\.(-?[0-9]+)\\.mca$")::matcher)
+                    .filter(Matcher::matches)
+                    .map(it -> new RegionPos(Integer.parseInt(it.group(1)), Integer.parseInt(it.group(2))))
+                    .flatMap(RegionPos::getContainedChunks)
+                    .collect(Collectors.toList());
+        }
+
+        AtomicInteger done = new AtomicInteger();
+        AtomicInteger total = new AtomicInteger(chunks.size());
+        progress.accept(done.get(), total.get());
+
+        StorageIoWorker io = (StorageIoWorker) getWorker();
+
+        // We ideally split the actual work of upgrading the chunk NBT across multiple threads, leaving a few for MC
+        int workThreads = Math.max(1, Runtime.getRuntime().availableProcessors() - 2);
+        ExecutorService workExecutor = Executors.newFixedThreadPool(workThreads, new DefaultThreadFactory("bobby-upgrade-worker", true));
+
+        try {
+            for (ChunkPos chunkPos : chunks) {
+                workExecutor.submit(() -> {
+                    NbtCompound nbt;
+                    try {
+                        nbt = io.getNbt(chunkPos);
+                    } catch (IOException e) {
+                        LOGGER.warn("Error reading chunk " + chunkPos.x + "/" + chunkPos.z + ":", e);
+                        nbt = null;
+                    }
+
+                    if (nbt == null) {
+                        progress.accept(done.get(), total.decrementAndGet());
+                        return;
+                    }
+
+                    nbt = updateChunkNbt(worldKey, null, nbt, generatorKey);
+
+                    io.setResult(chunkPos, nbt).join();
+
+                    progress.accept(done.incrementAndGet(), total.get());
+                });
+            }
+        } finally {
+            workExecutor.shutdown();
+        }
+
+        try {
+            //noinspection ResultOfMethodCallIgnored
+            workExecutor.awaitTermination(Long.MAX_VALUE, TimeUnit.NANOSECONDS);
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+
+        progress.accept(done.get(), total.get());
+    }
+
+    private record RegionPos(int x, int z) {
+        public Stream<ChunkPos> getContainedChunks() {
+            int baseX = x << 5;
+            int baseZ = z << 5;
+            ChunkPos[] result = new ChunkPos[32 * 32];
+            for (int x = 0; x < 32; x++) {
+                for (int z = 0; z < 32; z++) {
+                    result[x * 32 + z] = new ChunkPos(baseX + x, baseZ + z);
+                }
+            }
+            return Stream.of(result);
         }
     }